/alts enhancement
This adds the first feature mentioned in https://github.com/Warzone/TGM/issues/589

It gets the punishments for each alt and:
- Leaves the name white if no active mute/ban is found
- Makes the name yellow if only an active mute has been found
- Makes the name red if an active ban has been found.

Active bans override active mutes.

I'm not 100% sure if this is the most efficient way to add this feature, so it would be nice if someone with more knowledge could check the code.