Add Kit Layout Editor
Adds the ability of spectators to edit the kit layout if using an interactive chest item if:
- the map does not use the hardcoded kits from the GameClassModule
- the map does not have more than one kit (which applies to 97% of the maps in the rotation the time of commiting this)
- the kit editor module is not disabled

The kit editor module can be disabled globally using /kiteditor off in case there is an unexpected game-breaking bug. The normal kits will be given out instead. All edited kits will be deleted.

The code changes have been tested a lot. No duplication bugs or error throwings in the console were found.

The kits are only saved for the current ongoing map, since there is no way to save settings on a user level on the API yet. All edited kits will be unloaded in the cycling process.