Set variable even if no kits are present
If no kits were present before, such as on maps with classes, it would not change the isKitEditable variable, resulting in the kit editor item appearing in the spectator inventory, without the ability to open it as it would throw an error (kits are empty). This commit fixes it. Tested as well.