Broadcast /setnext to all players
If a map has been found after using /setnext, instead of only telling the sender the map has been set, it will instead by default broadcast it to everyone on the server.

The flag "-s" can be used to set the next map silently, resulting in only the sender and everyone with the permission to run the command to see the broadcast. If used, the flag must be added before the map name. Example: /sn -s Itty Bitty