Alts enhancements #3
Since recently, /punishments shows us a list of a user's punishments as well as all the punishments on his IP. This change made /alts mark alt account's names as yellow and red as well, while they might not be currently muted or banned, but have an alt which is.

This pulll request changes the following:
- Checks while iterating through the punishments of every alt, whether the punishment.getPunished() equals the alt name which is currently checked. If it does not, it continues with the next punishment.

- Checks the same while iterating through the punishments in the /punishments command, and doesn't list those punishments if the view is restricted (user does not have "tgm.punish.list")

- Checks for every alt in /alts if the player is online, and if so, marks the name green, just like in #668. It is overwritten by active mutes and bans.

Tested.