Add Block Interact Filter
- Cancels any interactions with blocks inside a region
- Includes block placement and breaking as well
- Reformats other filter types (I had to)

This is usefult to prevent the owner of a wool room to open their own wool/supply chests. Turf Wars is an example of a map where that is possible. Another solution would be to just build the wool room in a way which makes it impossible for the owner of the room to reach the chests, but that could be bypassed by breaking a block around the chest and using the short time until the block reappers to open the chest.