JSON Rotation Order Consistency (#763)
* JSON Rotation Order Consistency

In the reloadMaps() method, the JSON rotation map order previously lost its order. It would always use the order in which the maps have been loaded in the map library. I fixed it by first iterating through the map names instead of the map library.

This obviously only applies to rotations which are not shuffled.

Tested.

* Small changes